cleanup + CallInfo helpers
